+++ src/ts/beatmap/shared/parse.ts
@@ -3,14 +3,14 @@ import { CharacteristicOrder } from './characteristic';
 import logger from '../../logger';
 
 // deno-lint-ignore ban-types
-const tag = (func: Function) => {
-    return `[shared::parse::${func.name}]`;
+const tag = (name: string) => {
+    return `[shared::parse::${name}]`;
 };
 
 // TODO: more error check
 // TODO: contemplate whether to make pure function or keep as is
 export const info = (infoData: IInfoData): IInfoData => {
-    logger.info(tag(info), 'Parsing beatmap info v2.x.x');
+    logger.info(tag('info'), 'Parsing beatmap info v2.x.x');
     infoData._difficultyBeatmapSets.sort(
         (a, b) =>
             CharacteristicOrder[a._beatmapCharacteristicName] -
@@ -20,10 +20,10 @@ export const info = (infoData: IInfoData): IInfoData => {
         let num = 0;
         set._difficultyBeatmaps.forEach((a) => {
             if (a._difficultyRank - num <= 0) {
-                logger.warn(tag(info), a._difficulty + ' is unordered');
+                logger.warn(tag('info'), a._difficulty + ' is unordered');
             }
             if (DifficultyRank[a._difficulty] !== a._difficultyRank) {
-                logger.error(tag(info), a._difficulty + ' has invalid rank');
+                logger.error(tag('info'), a._difficulty + ' has invalid rank');
             }
             num = a._difficultyRank;
         });